requirements.txt
requests
python-dotenv
google-genai
spidev
pygame
opencv-python
flask
pandas
gpiod

